requirements.txt
anyio==4.8.0
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
mypy-extensions==1.0.0
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
PySocks==1.7.1
python-telegram-bot==21.10
PyWavelets==1.8.0
requests==2.32.3
schedule==1.2.2
scipy==1.15.1
selenium==4.27.1
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
